Refactor getBackground and getForeground
